Extract Windows1251 to its own file
