+++ escpos4k/src/commonTest/kotlin/cz/multiplatform/escpos4k/core/CommandBuilderTest.kt
@@ -2,6 +2,7 @@ package cz.multiplatform.escpos4k.core
 
 import cz.multiplatform.escpos4k.core.encoding.charset.Charset
 import cz.multiplatform.escpos4k.core.encoding.charset.IBM850
+import cz.multiplatform.escpos4k.core.encoding.charset.Windows1251
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldContainExactly
 
@@ -36,10 +37,10 @@ class CommandBuilderTest : FunSpec() {
               charset(Charset.default)
               charset(IBM850)
               charset(IBM850)
-              charset(Charset.Windows1251)
+              charset(Windows1251)
             }
         builder.commands shouldContainExactly
-            initSequence + Command.SelectCharset(IBM850) + Command.SelectCharset(Charset.Windows1251)
+            initSequence + Command.SelectCharset(IBM850) + Command.SelectCharset(Windows1251)
       }
     }
 
@@ -50,7 +51,7 @@ class CommandBuilderTest : FunSpec() {
               withCharset(Charset.default) {
                 withCharset(IBM850) {
                   withCharset(IBM850) {
-                    withCharset(Charset.Windows1251) {
+                    withCharset(Windows1251) {
                       text("innermost") //
                     }
                   }
@@ -61,8 +62,8 @@ class CommandBuilderTest : FunSpec() {
         builder.commands shouldContainExactly
             initSequence +
                 Command.SelectCharset(IBM850) +
-                Command.SelectCharset(Charset.Windows1251) +
-                Command.Text("innermost", Charset.Windows1251) +
+                Command.SelectCharset(Windows1251) +
+                Command.Text("innermost", Windows1251) +
                 Command.SelectCharset(IBM850) +
                 Command.SelectCharset(Charset.default)
       }
